fantasy-football/requirements.txt
tqdm
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.3
cssselect==1.0.3
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
lxml==4.2.3
parsel==1.5.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.7.1
pyOpenSSL==18.0.0
PyYAML==5.1.2
queuelib==1.5.0
requests==2.22.0
Scrapy==1.5.1
service-identity==17.0.0
six==1.11.0
Twisted==19.7.0
urllib3==1.25.3
w3lib==1.19.0
zope.interface==4.5.0
